Create helper_threshold.ino
This is a code that is used to generate real-time values of the solar panel sensor. We can see the values on the Serial monitor, for when the laser beam falls on the panel and when it does not. Based on this, we can manually approximate the threshold required in at the Receiver's code.